Finish implementing multi-chunk rendering; clean up foundations
The foundations of Crucible, while feature complete, need a lot of polish before we can start implementing more complicated features. Specifically, `Providers` and `RwLocks` don't interact as well as we'd want them to (we currently have to fall back on special-case macros to handle fetching ergonomically) and storages need to be greatly optimized (`world.new.rs` is the beginnings of my attempt at an archetypal ECS).