Remove `WorldAccessor` trait; make `Obj` follow external mutability
The `WorldAccessor` trait makes it way to easy to bloat the application size with monomorphizations of functions accepting a `World`.

Users should have to opt into `Rc` semantics for `Obj` because `Rc` semantics with dynamic extension add overhead to fetches in the form of atomic loads, which are expensive on some platforms. Furthermore, unique ownership of an `Obj` could allow users to use more efficient single-threaded borrowing checks. Users are now encouraged to use context objects to get `Obj` ancestry.