Start work on an alternative to the ECS
Right now, Geode's `oop` and `ecs` modules are pretty much isolated from
one another: users either choose the heavyweight `oop` object model or
the relatively lightweight `ecs` model. However, there's no reason we
can't combine both.

`geode-lab` is an experimental crate I'm working on that uses a multi-
arena garbage-collected `ObjectDB` to provide users with an allocator
that:

1. Provides *very* lightweight allocations.
2. Allows users to destroy objects at whim without having to wait for
   weak references to be destroyed.
3. Maintains spatial locality between objects with temporal locality.
4. Allows users to write single-threaded code that works in multi-
   threaded contexts through the use of thread exclusivity locks.
5. Allows users to change the arena in which an object is allocated on
   the fly.

The last point is especially relevant when comparing this model to the
ECS model it's replacing. If `GameObject` instances can effectively move
component `Obj` pointers into their proper arenas dynamically, we
effectively have a super opaque archetypal ECS. Combine that with a
`GameObjectSet` and all core functionality of an ECS is now available in
what appears to be an OOP object model.

This code is much messier and incomplete than the code in my other
commits. At this stage, I am still playing around with the interface
but am writing the basic scaffolding of the internals to make sure
things can actually be implemented. I'm probably going to create a
write-up of the implementation details before I take a stab at the final
product.